Refactor Terms of Service screen to use separate Compose component
* extract Compose UI logic from `TermsOfServiceFragment` into new `TermsOfServiceScreen`
* migrate `TermsOfServiceViewModel` to use `@HiltViewModel`
* remove manual ViewModel binding from `ViewModelModule`
* add Compose previews for `TermsOfServiceContent`
